# internal/app/delivery/http/user/dto/user.go
package dto

type UserResponse struct {
	Name     string `json:"name"`
	Username string `json:"username"`
	Email    string `json:"email"`
}

type UpdatePasswordRequest struct {
	OldPassword string `json:"old_password"`
	NewPassword string `json:"new_password"`
}

type UserUpdateInfoRequest struct {
	Name  string `json:"name"`
	Email string `json:"email"`
}